backend/requirements.txt
django
django-extensions
python-decouple
django-cors-headers
transformers
torch
djangorestframework